+++ hdlcc/tests/test_standalone.py
@@ -24,25 +24,16 @@ import re
 import subprocess as subp
 import shutil
 
+import mock
+
 from nose2.tools import such
 from nose2.tools.params import params
 
 _logger = logging.getLogger(__name__)
 
 HDLCC_LOCATION = p.join("hdlcc", "standalone.py")
-BUILDER_NAME = os.environ.get('BUILDER_NAME', None)
-BUILDER_PATH = os.environ.get("BUILDER_PATH", p.expanduser("~/ghdl/bin/"))
-
-_BUILDER_ENV = os.environ.copy()
-_BUILDER_ENV["PATH"] = p.expandvars(os.pathsep.join([BUILDER_PATH, \
-                                    _BUILDER_ENV["PATH"]]))
-
 TEST_SUPPORT_PATH = p.join(p.dirname(__file__), '..', '..', '.ci', 'test_support')
-
-if BUILDER_NAME is not None:
-    PROJECT_FILE = p.join(TEST_SUPPORT_PATH, "vim-hdl-examples", BUILDER_NAME + ".prj")
-else:
-    PROJECT_FILE = None
+VIM_HDL_EXAMPLES = p.abspath(p.join(TEST_SUPPORT_PATH, "vim-hdl-examples"))
 
 def shell(cmd):
     """Dummy wrapper for running shell commands, checking the return value and
@@ -53,7 +44,7 @@ def shell(cmd):
     stdout = []
     try:
         stdout += list(subp.check_output(
-            cmd, stderr=subp.STDOUT, env=_BUILDER_ENV).split("\n"))
+            cmd, stderr=subp.STDOUT).split("\n"))
     except subp.CalledProcessError as exc:
         stdout += list(exc.output.split("\n"))
 
@@ -67,7 +58,6 @@ def shell(cmd):
 
     if exc:
         _logger.warning("os.path: %s", os.environ["PATH"])
-        _logger.warning("_BUILDER_ENV path: %s", _BUILDER_ENV["PATH"])
         raise exc   # pylint: disable=raising-bad-type
 
     return stdout
@@ -75,58 +65,96 @@ def shell(cmd):
 with such.A("hdlcc standalone tool") as it:
     @it.has_setup
     def setup():
+        it.BUILDER_NAME = os.environ.get('BUILDER_NAME', None)
+        it.BUILDER_PATH = os.environ.get('BUILDER_PATH', None)
+        if it.BUILDER_NAME in ('ghdl', 'msim', 'xvhdl'):
+            it.PROJECT_FILE = p.join(VIM_HDL_EXAMPLES, it.BUILDER_NAME + '.prj')
+        else:
+            it.PROJECT_FILE = None
+
         it.assertTrue(p.exists(HDLCC_LOCATION))
+        if it.BUILDER_PATH:
+            it.patch = mock.patch.dict(
+                'os.environ',
+                {'PATH' : os.pathsep.join([it.BUILDER_PATH, os.environ['PATH']])})
+            it.patch.start()
 
     @it.has_teardown
     def teardown():
         cmd = ["coverage", "run",
-               HDLCC_LOCATION, PROJECT_FILE,
+               HDLCC_LOCATION, it.PROJECT_FILE,
                "--clean"]
 
-        shell(cmd)
-        if p.exists(p.join(p.dirname(PROJECT_FILE), '.build')):
-            shutil.rmtree(p.join(p.dirname(PROJECT_FILE), '.build'))
+        if it.PROJECT_FILE is not None:
+            shell(cmd)
+
+            if p.exists(p.join(p.dirname(it.PROJECT_FILE), '.build')):
+                shutil.rmtree(p.join(p.dirname(it.PROJECT_FILE), '.build'))
+
+        if p.exists(p.join(TEST_SUPPORT_PATH, 'vim-hdl-examples', '.build')):
+            shutil.rmtree(p.join(TEST_SUPPORT_PATH, 'vim-hdl-examples', '.build'))
+
+        if it.BUILDER_PATH:
+            it.patch.stop()
+
+        if p.exists('xvhdl.pb'):
+            os.remove('xvhdl.pb')
+        if p.exists('.xvhdl.init'):
+            os.remove('.xvhdl.init')
 
     with it.having("a valid project file"):
+
+        @it.has_setup
+        def setup():
+            cleanUp()
+
+        @it.has_teardown
+        def teardown():
+            cleanUp()
+
+        def cleanUp():
+            # Ensure there is no leftover files from previous runs
+            path = p.join(TEST_SUPPORT_PATH, 'vim-hdl-examples', '.build')
+            if p.exists(path):
+                shutil.rmtree(path)
+
 
         with it.having("a valid environment"):
 
             @it.should("print the tool's help")
             def test():
-                cmd = ["coverage", "run",
-                       HDLCC_LOCATION, "-h"]
-                shell(cmd)
-
-            @it.should("build a project")
-            def test():
-                cmd = ["coverage", "run",
-                       HDLCC_LOCATION, PROJECT_FILE,
-                       "--build"]
-
+                cmd = ['coverage', 'run',
+                       HDLCC_LOCATION, '-h']
                 shell(cmd)
 
             @it.should("run debug arguments")
             @params(
-                ("--debug-print-sources", ),
-                ("--debug-print-compile-order", ),
+                ('--debug-print-sources', ),
 
-                ("--build", "-s",
-                 p.join(TEST_SUPPORT_PATH, "vim-hdl-examples/another_library/foo.vhd")),
+                ('-vvv', '-s',
+                 p.join(TEST_SUPPORT_PATH, 'vim-hdl-examples',
+                        'another_library', 'foo.vhd')),
 
-                ("--debug-parse-source-file", "-s",
-                 p.join(TEST_SUPPORT_PATH, "vim-hdl-examples/another_library/foo.vhd")),
+                ('--debug-parse-source-file', '-s',
+                 p.join(TEST_SUPPORT_PATH, 'vim-hdl-examples',
+                        'another_library', 'foo.vhd')),
 
-                ("--debug-parse-source-file", "-s",
-                 p.join(TEST_SUPPORT_PATH, "vim-hdl-examples/basic_library/clk_en_generator.vhd")),
+                ('--debug-parse-source-file', '-s',
+                 p.join(TEST_SUPPORT_PATH, 'vim-hdl-examples', 'basic_library',
+                        'clk_en_generator.vhd')),
 
-                ("--debug-run-static-check", "-s",
-                 p.join(TEST_SUPPORT_PATH, "vim-hdl-examples/basic_library/clock_divider.vhd")),
+                ('--debug-run-static-check', '-s',
+                 p.join(TEST_SUPPORT_PATH, 'vim-hdl-examples',
+                        'another_library', 'foo.vhd')),
                 )
 
             def test(case, *args):
+                if not it.PROJECT_FILE:
+                    _logger.info("Test requires a project file")
+                    return
                 _logger.info("Running '%s'", case)
                 cmd = ["coverage", "run",
-                       HDLCC_LOCATION, PROJECT_FILE]
+                       HDLCC_LOCATION, it.PROJECT_FILE]
                 for arg in args:
                     cmd.append(arg)
 
@@ -135,25 +163,31 @@ with such.A("hdlcc standalone tool") as it:
 
             @it.should("save profiling info if requested")
             def test():
-                cmd = ["coverage", "run",
-                       HDLCC_LOCATION, PROJECT_FILE,
-                       "--debug-profiling", "output.stats"]
+                if not it.PROJECT_FILE:
+                    _logger.info("Test requires a project file")
+                    return
+                cmd = ['coverage', 'run',
+                       HDLCC_LOCATION, it.PROJECT_FILE,
+                       '--debug-profiling', 'output.stats']
 
-                if p.exists("output.stats"):
-                    os.remove("output.stats")
+                if p.exists('output.stats'):
+                    os.remove('output.stats')
 
                 shell(cmd)
 
-                it.assertTrue(p.exists("output.stats"))
-                os.remove("output.stats")
+                it.assertTrue(p.exists('output.stats'))
+                os.remove('output.stats')
 
             @it.should("control debugging level")
             def test():
-                cmd = ["coverage", "run",
-                       HDLCC_LOCATION, PROJECT_FILE,
-                       "--clean", "--build"]
+                if not it.PROJECT_FILE:
+                    _logger.info("Test requires a project file")
+                    return
+                cmd = ['coverage', 'run',
+                       HDLCC_LOCATION, it.PROJECT_FILE,
+                       '--clean', ]
 
-                it.assertEqual(shell(cmd), [""])
+                it.assertEqual(shell(cmd), [''])
 
                 previous = None
                 for level in range(1, 5):
@@ -161,6 +195,5 @@ with such.A("hdlcc standalone tool") as it:
                     it.assertTrue(len(stdout) >= previous)
                     previous = len(stdout)
 
-if BUILDER_NAME is not None:
-    it.createTests(globals())
+it.createTests(globals())
